requirements.txt
colorthief==0.2.1
favicon==0.7.0
Pillow==9.3.0
tldextract==3.4.0